% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrrogate_repos.R
\name{interrogate_repos}
\alias{interrogate_repos}
\title{Pull the curent number of R projects from CRAN, RForge, Bioconductor and GitHub.}
\usage{
interrogate_repos()
}
\value{
Named vector with the date and the number of packages in all 4 repos.
}
\description{
Using curl queries and parsing, the function pulls the current number of R projects from the 4 main repositories: CRAN, RForge, Bioconductor and GitHub.
}
\author{
Marcin Kierczak <\email{marcin.kierczak@scilifelab.se}>
}
