% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_repos_data.R
\name{plot_repos_data}
\alias{plot_repos_data}
\title{Plot the repositories data in glamorous way!}
\usage{
plot_repos_data(data)
}
\arguments{
\item{data}{-- long-format tibble as returned by the update_repos_data function.}
}
\value{
a ggplot2 plot object.
}
\description{
Plots the historical and the current data using tibble returned by the update_repos_data function.
Style of the plot is inspired by William Chase's lecture 'Glamour of graphics' at rstdio::rconf2020.
}
\author{
Marcin Kierczak <\email{marcin.kierczak@scilifelab.se}>
}
