% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_repos_data.R
\name{update_repos_data}
\alias{update_repos_data}
\title{Update the dataset with the number of available packages.}
\usage{
update_repos_data(path_to_dataset = NULL)
}
\arguments{
\item{path_to_dataset}{- specifies the path to the file where the historical data is stored.
If not given, internal package dataset will be used but it will not be updated.}
}
\value{
tibble (long-format) containing historical data and the new figures from today.
}
\description{
Queries all repos and pulls the current number of packages.
Then loads the old data and adds the new numbers to it. If the path to a dataset is specified,
the old data is loaded from that file and next, file is overwritten so that the new data is added.
Otherwise the built-in dataset is used as the historical data and the new data will not be saved.
}
\author{
Marcin Kierczak <\email{marcin.kierczak@scilifelab.se}>
}
